backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pymongo==4.6.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic[email]==2.4.2
bcrypt==4.0.1
python-jose==3.3.0 